Moved target test in standalone file
Added check for wrapper presence
